feat: make get size by breakpoint shared util
