+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './color'
 export * from './convert-case-style'
 export * from './page'
+export * from './size'
