+++ src/types/size.ts
@@ -1,0 +1,5 @@
+import type { Breakpoint } from '~/styles'
+
+type SizeValue = number
+
+export type Size = SizeValue | { [key in Breakpoint]?: SizeValue }
